refactor(phase-6g): slice render.php (2,525 LOC) into includes/render/*
The render layer's god-module is now a thin umbrella loader (29
LOC) plus 5 focused files under includes/render/. Each file owns
one phase of the request lifecycle:

- includes/render/body-classes.php (49 LOC) —
    desktop_mode_admin_body_classes (admin_body_class filter)
- includes/render/assets.php (309 LOC) —
    desktop_mode_enqueue_assets (admin_enqueue_scripts hook)
- includes/render/shell.php (87 LOC) —
    desktop_mode_render_shell (in_admin_header @ 5)
- includes/render/chromeless-bridge.php (2,054 LOC) —
    desktop_mode_chromeless_offset_neutralizer_script +
    desktop_mode_chromeless_bridge_script (the heavyweight, 95%
    of the file is the inline JS string the iframe runs)
- includes/render/classic-link-interceptor.php (117 LOC) —
    desktop_mode_classic_link_interceptor (admin_footer)

`includes/render.php` is now an umbrella `require_once` of the
five files so plugins or test bootstraps that
`require_once 'includes/render.php'` continue to load the same
hooks they always did.

Behaviour unchanged. PHPUnit confirmed via docker exec on the
running wordpress-alcazaba dev container: 573 tests / 1,300
assertions still green after the move.

Phase 6 totals so far:
- helpers.php:       1,609 → 153   LOC (-1,456)
- components.php:    2,101 → 376   LOC (-1,725)
- render.php:        2,525 → 29    LOC (-2,496)
                     -----   ----
                     6,235   558    (-5,677 LOC across 3 files)
